+++ src/openvpn/shaper.h
@@ -147,10 +147,10 @@ shaper_wrote_bytes(struct shaper *s, int nbytes)
         tv_add(&s->wakeup, &tv);
 
 #ifdef SHAPER_DEBUG
-        dmsg(D_SHAPER_DEBUG, "SHAPER shaper_wrote_bytes bytes=%d delay=%ld sec=%lld usec=%ld",
+        dmsg(D_SHAPER_DEBUG, "SHAPER shaper_wrote_bytes bytes=%d delay=%ld sec=%"PRIi64" usec=%ld",
              nbytes,
              (long)tv.tv_usec,
-             (long long)s->wakeup.tv_sec,
+             (int64_t)s->wakeup.tv_sec,
              (long)s->wakeup.tv_usec);
 #endif
     }
